dom/canvas/test/test_offscreencanvas_basic_operations.html

<!DOCTYPE HTML>
<html>
<head>
<title>WebGL in OffscreenCanvas</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
</head>
<body>
<canvas id="c-ref" width="64" height="64"></canvas>
<script>

SimpleTest.waitForExplicitFinish();

function createCanvas() {
  var htmlCanvas = document.createElement('canvas');
  htmlCanvas.width = 64;
  htmlCanvas.height = 64;
  document.body.appendChild(htmlCanvas);
  return htmlCanvas;
}

function testToDataURL(canvas, test) {
  // testing toDataURL
  // Fill c-ref with green color.
  var c = document.getElementById("c-ref");
  var ctx = c.getContext("2d");
  ctx.rect(0, 0, 64, 64);
  ctx.fillStyle = "#00FF00";
  ctx.fill();
  ok(c.toDataURL() == canvas.toDataURL(), "[" + test + "]" +
     "toDataURL should return a 64x64 green square");
}

function runTest() {
  var stillRunning = 0;

  var startWorker = function(canvas, canvasType) {
    stillRunning++;
    var worker = new Worker("offscreencanvas.js");

    ok(canvas, "[" + canvasType + "]" + "Should have HTML canvas element");
    ok(worker, "[" + canvasType + "]" + "Web worker successfully created");

    worker.onmessage = function(evt) {
      var msg = evt.data || {};
      if (msg.type == "test") {
        ok(msg.result, "[" + canvasType + "]" + msg.name);
      }
      if (msg.type == "finish") {
        testToDataURL(canvas, canvasType);
        worker.terminate();
        if (--stillRunning == 0) {
          SimpleTest.finish();
        }
      }
    }

    ok(canvas.transferControlToOffscreen, "[" + canvasType + "]" +
       "HTMLCanvasElement has transferControlToOffscreen function");

    var offscreenCanvas = canvas.transferControlToOffscreen();
    ok(offscreenCanvas, "[" + canvasType + "]" +
       "Expected transferControlToOffscreen to succeed");

    worker.postMessage({canvasType: canvasType, testType: 'basic',
                        canvas: offscreenCanvas}, [offscreenCanvas]);
  }

  startWorker(createCanvas(), '2d');
  startWorker(createCanvas(), 'webgl');
}

SpecialPowers.pushPrefEnv({'set': [
  ['gfx.offscreencanvas.enabled', true],
  ['webgl.force-enabled', true],
]}, runTest);

</script>
</body>
</html>
